+++ src/components/button-postcode/button-postcode.styled.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+import { Button } from '../button-normal/button-normal.styled';
+
+// DESC: 버튼 스타일링(현재는 기존의 Button 스타일 상속)
+export const PostcodeButton = styled(Button)``;
